feat(meter): add subject display name to csv export
