+++ openmeter/meter/httphandler/handler.go
@@ -8,6 +8,7 @@ import (
 	"github.com/openmeterio/openmeter/openmeter/meter"
 	"github.com/openmeterio/openmeter/openmeter/namespace/namespacedriver"
 	"github.com/openmeterio/openmeter/openmeter/streaming"
+	"github.com/openmeterio/openmeter/openmeter/subject"
 	"github.com/openmeterio/openmeter/pkg/framework/commonhttp"
 	"github.com/openmeterio/openmeter/pkg/framework/transport/httptransport"
 )
@@ -35,6 +36,7 @@ type handler struct {
 	options          []httptransport.HandlerOption
 	meterService     meter.ManageService
 	streaming        streaming.Connector
+	subjectService   subject.Service
 }
 
 func (h *handler) resolveNamespace(ctx context.Context) (string, error) {
@@ -50,6 +52,7 @@ func New(
 	namespaceDecoder namespacedriver.NamespaceDecoder,
 	meterService meter.ManageService,
 	streaming streaming.Connector,
+	subjectService subject.Service,
 	options ...httptransport.HandlerOption,
 ) Handler {
 	return &handler{
@@ -57,5 +60,6 @@ func New(
 		options:          options,
 		meterService:     meterService,
 		streaming:        streaming,
+		subjectService:   subjectService,
 	}
 }
