+++ openmeter/meter/httphandler/query.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"time"
 
 	"github.com/openmeterio/openmeter/api"
 	"github.com/openmeterio/openmeter/openmeter/meter"
@@ -130,125 +129,6 @@ func (h *handler) QueryMeter() QueryMeterHandler {
 			httptransport.WithOperationName("queryMeter"),
 		)...,
 	)
-}
-
-type (
-	QueryMeterCSVParams   = QueryMeterParams
-	QueryMeterCSVRequest  = QueryMeterRequest
-	QueryMeterCSVResponse = commonhttp.CSVResponse
-	QueryMeterCSVHandler  httptransport.HandlerWithArgs[QueryMeterCSVRequest, QueryMeterCSVResponse, QueryMeterCSVParams]
-)
-
-// QueryMeterCSV returns a handler for query meter.
-func (h *handler) QueryMeterCSV() QueryMeterCSVHandler {
-	return httptransport.NewHandlerWithArgs(
-		func(ctx context.Context, r *http.Request, params QueryMeterCSVParams) (QueryMeterCSVRequest, error) {
-			ns, err := h.resolveNamespace(ctx)
-			if err != nil {
-				return QueryMeterCSVRequest{}, err
-			}
-
-			return QueryMeterCSVRequest{
-				namespace: ns,
-				idOrSlug:  params.IdOrSlug,
-				params:    params.QueryMeterParams,
-			}, nil
-		},
-		func(ctx context.Context, request QueryMeterCSVRequest) (QueryMeterCSVResponse, error) {
-			meter, err := h.meterService.GetMeterByIDOrSlug(ctx, meter.GetMeterInput{
-				Namespace: request.namespace,
-				IDOrSlug:  request.idOrSlug,
-			})
-			if err != nil {
-				return nil, fmt.Errorf("failed to get meter: %w", err)
-			}
-
-			params, err := ToQueryParamsFromAPIParams(meter, request.params)
-			if err != nil {
-				return nil, fmt.Errorf("failed to construct query meter params: %w", err)
-			}
-
-			rows, err := h.streaming.QueryMeter(ctx, request.namespace, meter, params)
-			if err != nil {
-				return nil, fmt.Errorf("failed to query meter: %w", err)
-			}
-
-			response := NewQueryMeterCSVResult(meter.Key, params.GroupBy, rows)
-
-			return response, nil
-		},
-		commonhttp.CSVResponseEncoder[QueryMeterCSVResponse],
-		httptransport.AppendOptions(
-			h.options,
-			httptransport.WithOperationName("queryMeterCSV"),
-		)...,
-	)
-}
-
-// QueryMeterCSVResult is a CSV response for query meter.
-var _ QueryMeterCSVResponse = (*queryMeterCSVResult)(nil)
-
-func NewQueryMeterCSVResult(meterSlug string, queryGroupBy []string, rows []meter.MeterQueryRow) QueryMeterCSVResponse {
-	return &queryMeterCSVResult{
-		meterSlug:    meterSlug,
-		queryGroupBy: queryGroupBy,
-		rows:         rows,
-	}
-}
-
-type queryMeterCSVResult struct {
-	meterSlug    string
-	queryGroupBy []string
-	rows         []meter.MeterQueryRow
-}
-
-// Records returns the CSV records.
-func (a *queryMeterCSVResult) Records() [][]string {
-	records := [][]string{}
-
-	// Filter out the subject from the group by keys
-	groupByKeys := []string{}
-	for _, k := range a.queryGroupBy {
-		if k == "subject" {
-			continue
-		}
-		groupByKeys = append(groupByKeys, k)
-	}
-
-	// CSV headers
-	headers := []string{"window_start", "window_end", "subject"}
-	if len(groupByKeys) > 0 {
-		headers = append(headers, groupByKeys...)
-	}
-	headers = append(headers, "value")
-	records = append(records, headers)
-
-	// CSV data
-	for _, row := range a.rows {
-		data := []string{row.WindowStart.Format(time.RFC3339), row.WindowEnd.Format(time.RFC3339)}
-		if row.Subject != nil {
-			data = append(data, *row.Subject)
-		} else {
-			data = append(data, "")
-		}
-		for _, k := range groupByKeys {
-			var groupByValue string
-
-			if row.GroupBy[k] != nil {
-				groupByValue = *row.GroupBy[k]
-			}
-			data = append(data, groupByValue)
-		}
-		data = append(data, fmt.Sprintf("%f", row.Value))
-		records = append(records, data)
-	}
-
-	return records
-}
-
-// FileName returns the CSV file name.
-func (a *queryMeterCSVResult) FileName() string {
-	return fmt.Sprintf("%s.csv", a.meterSlug)
 }
 
 type QueryMeterPostRequest struct {
